data seed, some entity changes

+++ src/Services/Main/Main.Application/Models/Invoice.cs
@@ -5,7 +5,7 @@ namespace Main.Application.Models;
 [Owned]
 public class Invoice
 {
-    public double Sum { get; }
+    public decimal Sum { get; }
     public DateTime OperationTime { get; }
 
     // ReSharper disable once UnusedMember.Local
@@ -13,7 +13,7 @@ public class Invoice
     {
     }
 
-    public Invoice(DateTime operationTime, double sum)
+    public Invoice(DateTime operationTime, decimal sum)
     {
         Sum = sum;
         OperationTime = operationTime;
